Got the basic api documentation function to work
