+++ SolverManager/src/main/java/com/oops/solvermanager/Requests/ProblemRequest.java
@@ -19,31 +19,4 @@ public class ProblemRequest {
     public SolverBody[] getSolversToUse() {
         return solversToUse;
     }
-
-    private class SolverBody {
-        private String solverName;
-        private int numberVCPU;
-        private int maxMemory; // Max Memory in MB
-        private int timeout; // Seconds to timeout
-
-        SolverBody() {
-        }
-
-        public int getMaxMemory() {
-            return maxMemory;
-        }
-
-        public int getNumberVCPU() {
-            return numberVCPU;
-        }
-
-        public String getSolverName() {
-            return solverName;
-        }
-
-        public int getTimeout() {
-            return timeout;
-        }
-
-    }
 }
